+++ deepvision/__init__.py
@@ -1,5 +1,3 @@
-from deepvision import datasets
-from deepvision import evaluation
-from deepvision import models
+from deepvision import datasets, evaluation, models
 
 __version__ = "0.1.6"
