+++ deepvision/layers/relative_positional_transformer_encoder.py
@@ -14,9 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Optional
-from typing import Tuple
-from typing import Type
+from typing import Optional, Tuple, Type
 
 import torch
 import torch.nn as nn
@@ -71,12 +69,18 @@ class RelativePositionalTransformerEncoder(nn.Module):
 
         self.norm2 = norm_layer(dim)
 
+        """
         self.mlp = MLP(
             output_dim=dim,
             embed_dim=int(dim * mlp_ratio),
             activation=act_layer,
             num_layers=2,
             backend="pytorch",
+        )
+        """
+
+        self.mlp = MLPBlock(
+            embedding_dim=dim, mlp_dim=int(dim * mlp_ratio), act=act_layer
         )
 
         self.window_size = window_size
@@ -98,3 +102,19 @@ class RelativePositionalTransformerEncoder(nn.Module):
         x = x + self.mlp(self.norm2(x))
 
         return x
+
+
+class MLPBlock(nn.Module):
+    def __init__(
+        self,
+        embedding_dim: int,
+        mlp_dim: int,
+        act: Type[nn.Module] = nn.GELU,
+    ) -> None:
+        super().__init__()
+        self.lin1 = nn.Linear(embedding_dim, mlp_dim)
+        self.lin2 = nn.Linear(mlp_dim, embedding_dim)
+        self.act = act()
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self.lin2(self.act(self.lin1(x)))
