+++ deepvision/layers/sam_mask_decoder.py
@@ -14,16 +14,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List
-from typing import Tuple
-from typing import Type
+from typing import List, Tuple, Type
 
 import torch
 from torch import nn
 from torch.nn import functional as F
 
 from deepvision.layers.layernorm2d import LayerNorm2d
-from deepvision.layers.mlp import MLP
+
+# from deepvision.layers.mlp import MLP
 
 
 class MaskDecoder(nn.Module):
@@ -74,6 +73,19 @@ class MaskDecoder(nn.Module):
             ),
             activation(),
         )
+
+        self.output_hypernetworks_mlps = nn.ModuleList(
+            [
+                MLP(transformer_dim, transformer_dim, transformer_dim // 8, 3)
+                for i in range(self.num_mask_tokens)
+            ]
+        )
+
+        self.iou_prediction_head = MLP(
+            transformer_dim, iou_head_hidden_dim, self.num_mask_tokens, iou_head_depth
+        )
+
+        """
         self.output_hypernetworks_mlps = nn.ModuleList(
             [
                 MLP(
@@ -96,6 +108,7 @@ class MaskDecoder(nn.Module):
             num_layers=iou_head_depth,
             backend="pytorch",
         )
+        """
 
     def forward(
         self,
@@ -182,3 +195,28 @@ class MaskDecoder(nn.Module):
         iou_pred = self.iou_prediction_head(iou_token_out)
 
         return masks, iou_pred
+
+
+class MLP(nn.Module):
+    def __init__(
+        self,
+        input_dim: int,
+        hidden_dim: int,
+        output_dim: int,
+        num_layers: int,
+        sigmoid_output: bool = False,
+    ) -> None:
+        super().__init__()
+        self.num_layers = num_layers
+        h = [hidden_dim] * (num_layers - 1)
+        self.layers = nn.ModuleList(
+            nn.Linear(n, k) for n, k in zip([input_dim] + h, h + [output_dim])
+        )
+        self.sigmoid_output = sigmoid_output
+
+    def forward(self, x):
+        for i, layer in enumerate(self.layers):
+            x = F.relu(layer(x)) if i < self.num_layers - 1 else layer(x)
+        if self.sigmoid_output:
+            x = F.sigmoid(x)
+        return x
