+++ deepvision/models/__init__.py
@@ -1,52 +1,54 @@
 from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B0,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B1,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B2,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B3,
+    EfficientNetV2L,
+    EfficientNetV2M,
+    EfficientNetV2S,
 )
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2L
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2M
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2S
-from deepvision.models.classification.mix_transformer.mit import MiTB0
-from deepvision.models.classification.mix_transformer.mit import MiTB1
-from deepvision.models.classification.mix_transformer.mit import MiTB2
-from deepvision.models.classification.mix_transformer.mit import MiTB3
-from deepvision.models.classification.mix_transformer.mit import MiTB4
-from deepvision.models.classification.mix_transformer.mit import MiTB5
-from deepvision.models.classification.resnet.resnetv2 import ResNet18V2
-from deepvision.models.classification.resnet.resnetv2 import ResNet34V2
-from deepvision.models.classification.resnet.resnetv2 import ResNet50V2
-from deepvision.models.classification.resnet.resnetv2 import ResNet101V2
-from deepvision.models.classification.resnet.resnetv2 import ResNet152V2
-from deepvision.models.classification.vision_transformer.vit import ViTB16
-from deepvision.models.classification.vision_transformer.vit import ViTB32
-from deepvision.models.classification.vision_transformer.vit import ViTH16
-from deepvision.models.classification.vision_transformer.vit import ViTH32
-from deepvision.models.classification.vision_transformer.vit import ViTL16
-from deepvision.models.classification.vision_transformer.vit import ViTL32
-from deepvision.models.classification.vision_transformer.vit import ViTS16
-from deepvision.models.classification.vision_transformer.vit import ViTS32
-from deepvision.models.classification.vision_transformer.vit import ViTTiny16
-from deepvision.models.classification.vision_transformer.vit import ViTTiny32
-from deepvision.models.classification.vision_transformer_detector.vit_det_pt import (
-    ViTDetBackbone,
+from deepvision.models.classification.mix_transformer.mit import (
+    MiTB0,
+    MiTB1,
+    MiTB2,
+    MiTB3,
+    MiTB4,
+    MiTB5,
 )
-from deepvision.models.segmentation.sam.sam import Sam
-from deepvision.models.segmentation.segformer.segformer import SegFormerB0
-from deepvision.models.segmentation.segformer.segformer import SegFormerB1
-from deepvision.models.segmentation.segformer.segformer import SegFormerB2
-from deepvision.models.segmentation.segformer.segformer import SegFormerB3
-from deepvision.models.segmentation.segformer.segformer import SegFormerB4
-from deepvision.models.segmentation.segformer.segformer import SegFormerB5
+from deepvision.models.classification.resnet.resnetv2 import (
+    ResNet18V2,
+    ResNet34V2,
+    ResNet50V2,
+    ResNet101V2,
+    ResNet152V2,
+)
+from deepvision.models.classification.vision_transformer.vit import (
+    ViTB16,
+    ViTB32,
+    ViTH16,
+    ViTH32,
+    ViTL16,
+    ViTL32,
+    ViTS16,
+    ViTS32,
+    ViTTiny16,
+    ViTTiny32,
+)
+from deepvision.models.classification.vision_transformer_detector.vit_det import ViTDetB
+from deepvision.models.segmentation.sam.sam import SAM_B
+from deepvision.models.segmentation.segformer.segformer import (
+    SegFormerB0,
+    SegFormerB1,
+    SegFormerB2,
+    SegFormerB3,
+    SegFormerB4,
+    SegFormerB5,
+)
 from deepvision.models.volumetric import volumetric_utils
-from deepvision.models.volumetric.nerf.nerf import NeRF
-from deepvision.models.volumetric.nerf.nerf import NeRFLarge
-from deepvision.models.volumetric.nerf.nerf import NeRFMedium
-from deepvision.models.volumetric.nerf.nerf import NeRFSmall
-from deepvision.models.volumetric.nerf.nerf import NeRFTiny
+from deepvision.models.volumetric.nerf.nerf import (
+    NeRF,
+    NeRFLarge,
+    NeRFMedium,
+    NeRFSmall,
+    NeRFTiny,
+)
