+++ deepvision/models/classification/efficientnet/efficientnet_weight_mapper.py
@@ -17,27 +17,18 @@ import os
 import tensorflow as tf
 import torch
 
-from deepvision.layers import fused_mbconv
-from deepvision.layers import mbconv
-from deepvision.layers.fused_mbconv import __FusedMBConvPT
-from deepvision.layers.fused_mbconv import __FusedMBConvTF
-from deepvision.layers.mbconv import __MBConvPT
-from deepvision.layers.mbconv import __MBConvTF
+from deepvision.layers import fused_mbconv, mbconv
+from deepvision.layers.fused_mbconv import __FusedMBConvPT, __FusedMBConvTF
+from deepvision.layers.mbconv import __MBConvPT, __MBConvTF
 from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B0,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B1,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B2,
-)
-from deepvision.models.classification.efficientnet.efficientnetv2 import (
     EfficientNetV2B3,
+    EfficientNetV2L,
+    EfficientNetV2M,
+    EfficientNetV2S,
 )
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2L
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2M
-from deepvision.models.classification.efficientnet.efficientnetv2 import EfficientNetV2S
 from deepvision.models.classification.efficientnet.efficientnetv2_pt import (
     EfficientNetV2PT,
 )
