+++ deepvision/models/classification/efficientnet/efficientnetv2_tf.py
@@ -17,8 +17,7 @@ import math
 import tensorflow as tf
 from tensorflow.keras import layers
 
-from deepvision.layers import FusedMBConv
-from deepvision.layers import MBConv
+from deepvision.layers import FusedMBConv, MBConv
 from deepvision.utils.utils import parse_model_inputs
